Extract shared AbstractPostgresDigestPersistenceAdapter
digest-pg and digest-pglite duplicated ~85-90% identical SQL for the
digest persistence methods. Move the pure-SQL logic (upsert, delete,
read, query, notify) into a single AbstractPostgresDigestPersistenceAdapter
in digest-persistence, written once against $N positional placeholders
over a minimal PostgresQueryRunner driver shim. Each adapter now only
supplies connect/disconnect/listen plumbing and a runQuery hook that
adapts its driver's query() result shape.

digest-pg no longer needs pg-bind now that its SQL is positional, so
the dependency and its type shim are removed.
