+++ .changeset/digest-postgres-base.md
@@ -1,0 +1,7 @@
+---
+"@telia-oss/xjog-digest-persistence": minor
+"@telia-oss/xjog-digest-pg": minor
+"@telia-oss/xjog-digest-pglite": minor
+---
+
+Extract a shared `AbstractPostgresDigestPersistenceAdapter` (positional `$N` SQL over a `PostgresQueryRunner` shim); digest-pg/digest-pglite now provide only the driver layer.
