+++ packages/digest-persistence/src/PostgresDigestRow.ts
@@ -1,0 +1,12 @@
+/**
+ * Digest record row directly from the SQL query, shared by every Postgres
+ * compatible driver (`pg`, PGlite, …).
+ */
+export type PostgresDigestRow = {
+  created: number;
+  timestamp: number;
+  machineId: string;
+  chartId: string;
+  key: string;
+  value: string;
+};
