+++ packages/digest-persistence/src/index.ts
@@ -1,5 +1,8 @@
+export * from './AbstractPostgresDigestPersistenceAdapter';
 export * from './DigestEntries';
 export * from './DigestEntry';
 export * from './DigestPersistenceAdapter';
 export * from './DigestQuery';
 export * from './filterQuery';
+export * from './PostgresDigestRow';
+export * from './PostgresQueryRunner';
