+++ packages/digest-pg/src/PostgresDigestRow.ts
@@ -1,11 +1,1 @@
-/**
- * Digest record row directly from the SQL query
- */
-export type PostgresDigestRow = {
-  created: number;
-  timestamp: number;
-  machineId: string;
-  chartId: string;
-  key: string;
-  value: string;
-};
+export type { PostgresDigestRow } from '@telia-oss/xjog-digest-persistence';
