+++ packages/digest-pglite/src/PGliteDigestRow.ts
@@ -1,11 +1,1 @@
-/**
- * Digest record row directly from the SQL query
- */
-export type PGliteDigestRow = {
-  created: number;
-  timestamp: number;
-  machineId: string;
-  chartId: string;
-  key: string;
-  value: string;
-};
+export type { PostgresDigestRow as PGliteDigestRow } from '@telia-oss/xjog-digest-persistence';
